Refactor SharedTreeCore and create SharedTreeView (#15613)
## Description

This PR consolidates the various implementations of `ISharedTreeView`
into a single canonical implementation, now contained in a class called
`SharedTreeView`. This class is composed over by `SharedTree` and is
also used as the implementation for forks of views (i.e. it replaces
`SharedTreeFork`). It can also be used in a vacuum - thus, we can use it
to unit test views, indexes, invalidation, etc. without having to
connect to a mock fluid runtime or any Fluid modules at all.

To accomplish this, the control relationship between `SharedTreeCore`
and the local branch has been inverted. `SharedTreeCore` now responds to
changes on the local branch via events which are emitted by the local
branch. I had attempted this approach in the past but had trouble
defining the right set of events, however, this PR updates
`SharedTreeBranchEvents` to be powerful enough for the various scenarios
we care about to be detected while still being quite generic. This
inversion of control means a lot of boilerplate has been removed from
`SharedTreeCore`.

The `SharedTreeView` class is exposed directly rather than using an
interface to reduce its API. This means that we no longer need to do
`instanceof` checks and casts in order to extract the branch from an
`ISharedTreeView` in order to do rebase and merge operations. This is
made possible by a new `rebase` method on `ISharedTreeView` which is
symmetrical to `rebaseOnto` (which still only exists on forks, not on
the root `SharedTree`).

## Breaking Changes

The `ISharedTreeFork` interface has been removed in favor of
`SharedTreeView`.