(protocol-base): move scribeHelper to services-client (#15424)
The functionality related to scribe, blobs, quorum, and some git
functions can be found in the `@fluidframework/protocol-base` package.

The `scribeHelper` is only used in the server release group. I have
moved `scribeHelper` to the `@fluidframework/server-services-client`
package. Ideally, the `scribeHelper` should be placed in
`@fluidframework/server-lambdas#scribe`. However, due to cyclic
workspace dependencies between lambdas and test-utils, this arrangement
is not feasible.

The end goal is to: either move the protocol-base package to the client
release group or eliminate the package entirely by appropriately
relocating the interfaces, functions, and classes within the codebase.