Update client deps (#15579)
## Description

Update lock file with newer dependency versions. Other changes are
needed to keep everything working.

This was done by deleting and regenerating the lock file and
node-modules, then debugging the various issues that occurred.
This was repeated until no more issues occurred.

After this change, deleting the root lock file and ./node-modules and
regenerating them becomes a no-op.

Note that this does not change anything that is shipped to Fluid
Framework users: all changes are in the root package json, examples and
the lock file, with one exception:
experimental/PropertyDDS/packages/property-inspector-table/package.json
had a dep pinned to avoid a build error. This will fix the issue for
anyone installing this package as well.