Remove danger dependency from root - part 1 (#15593)
## Description

`danger` was added as a dependency in the root `package.json` as a
stopgap solution when upgrading to Node 16 due to how `RunBundlestats`
works. This is the followup to remove it **(ADO:3710)**. First, this
needs to be added as a dependency in `build-cli`. After releasing the
new version of the cli, we can remove the dependency from the root
`package.json` file, as `danger` will be brought by `build-cli`.