Minimal working node existence constraint (#15594)
## Description
This is a minimally working node existence constraint. This PR
explicitly doesn't handle the revive case, which will be done as a
follow-up as I wanted this PR to be used for discussion about a few
outstanding questions I have regarding repair data.

This implementation works by adding a deleted parameter to
`rebaseNodeChange`, which is only true in a single case right now and is
an optional parameter. This is using a "notification" pattern rather
than signifying whether or not the node currently exists in the input
context (as in deleted will only be true when the node is being
deleted).

I'm mostly concerned about hooking up the repair store. Without hooking
it up, we get "DummyRevivedValue" in the test I wrote. I tried hooking
up a singular RepairDataStore and capturing each local change, but was
getting errors and switched to the way the UndoRedoManager does it and
am maintaining a map of `<RevisionTag, RepairDataStore>` even though it
seems like I should be able to just have a single `RepairDataStore` and
call capture on each commit. If the map is necessary, when is it ok to
evict? As this is currently implemented, the map will grow unbounded.