Adding config.json file for E2E tests and moving time bench to custom (#15644)
## Description
This commit addresses 2 outstanding issues: 
1) Move the E2E Load and Summarizer run scenarios to have an optipn of
being populated from a config json file. This should allow consumers to
change and test their scenarios without recompiling the code
2) In order to optimize the time it takes to run the performance tests,
we are moving towards the new lower level and more flexible benchmarking
API with manual timing, that Craig created a few weeks back. That should
address some of the timeouts we are experiencing during the performance
executions on the pipeline.