Add suggested commands (#15651)
From @alexvy86:
> Oh, also I think we'll want both versions of the scripts, with and
without :coverage. The pipeline step that calls them is on an if-else
block that depends on a variable we can change when kicking off a run
(and defaults to "run the with-coverage" versions, IIRC). So we can
technically request that it calls the script names without the :coverage
suffix. That's why all the ones around these new ones come in both
flavors.

This PR adds those commands as the first one was to fix the internal
build.