[GC] Refactor the GC functions to clearly distinguish mark and sweep phases (#15598)
## Description
This change refactors the code around the main garbage collection run by
separating it into 3 steps:
1. Pre GC step - This includes initializing from base state (if not
already done) and notifying container runtime that GC is starting.
2. GC step - This includes 2 main parts - running mark phase and running
sweep phase.
3. Post GC step - This includes clearing state and logging any events.

The goal is to clearly indicate what logic runs in which step and also
consolidate the logic of mark and sweep phases.

## Reviewer Guidance
This is a refactor. There is no change in functionality.