Removing the dependency of aqueduct on DependencyContainer (#15391)
## Description
Removing the dependency of aqueduct on a particular DependencyContainer
implementation. This allows flexibility for how to implement the
IFluidDependencySynthesizer.

This PR doesn't quite remove all dependencies to avoid a breaking
change. It does however remove the dependency of pureDataObjectFastory
on DependencyContainer which is what matters the most.

## Breaking Changes
IContainerRuntime is no longer registered in DependencyContainer in
BaseContainerRuntimeFactory. This is not an interface breaking change.

## Reviewer Guidance
- Any concerns here about backwards compatibility?
IProvideContainerRuntime has been marked deprecated for 8 months.
- I am not sure why I can't assign {} to
AsyncFluidObjectProvider<I["OptionalProviders"]> requiring a cast. Help
required from someone more familiar with those types.

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>