Fixed incrementalSummaryViolation error logged due to GC state updated DS count reset (#15663)
## Bug
This error is logged when the number of datas store summarized in a
summary is greater than ops since last summary + number of data stores
whose GC state updated.
Basically, `summarizedDataStoreCount <= opsSinceLastSummary +
gcStateUpdatedDataStoreCount`.

GC updates its state on every GC run / summarize, i.e,
`gcStateUpdatedDataStoreCount` will be reset when GC ran and then the
summary failed.
Whereas, `summarizedDataStoreCount` and `opsSinceLastSummary` are
updated on every successful summary, i.e., when a summary is ack'd.
So, when a summary fails, GC has already updated its state whereas the
other two states are not updated. So, we may see this error logged in
the next summary if `gcStateUpdatedDataStoreCount` is non-zero.

## Fix
The fix is that GC tracks `gcStateUpdatedDataStoreCount` across
successful summaries and exposes it in the `IGarbageCollector`
interface. This is okay because the interace is internal (not public).

## Alternatives considered
- Another approach was to add a new property to `IGCStats` called
`gcStateUpdatedDataStoreCountSinceLastSummary` and use that. But that
would make the stats inconsistent - Currently, they all indicate stats
between GC runs (and not successful summary runs). It also makes this a
breaking change because `IGCStats` is public (this isn't a big deal but
contributed to the decision).
- I also considered fixing this in `SummaryGenerator` by tracking the
count across summary attempts. So, on failure, it would cache the count
and use it in the next attempt. But, the problem is attemps are managed
by `RunningSummarizer` and so, it would have to track that which would
make the code very complicated. Moreover, GC code seems like the right
place for this.

Note: With this change, `IRootSummaryTreeWithStats` is no longer
required. I didn't change that in this PR because it's a breaking change
-
[AB#4421](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4421)


[AB#4260](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4260)