Add unit tests for the DDS fuzz harness (#15529)
## Description

This adds a suite of unit tests to the DDS fuzz harness exported by
`@fluid-internal/test-dds-utils`. I exported all of the mixin
functionality at the file level but not the package level, as that was
frequently reasonable to test without the rest of the infrastructure.

To test the top-level suite infrastructure, I wrote a few test suites
which invoke mocha on other files. This strategy allows verifying
user-driven debugging behavior like ".only can be applied to the suite"
or "files with the list of attempted operations get written on failure."

I didn't test any super nitty-gritty details, but this should prevent us
from regressing a lot of basic issues that could occur.