Update blobs cache when getting a whole snapshot in the wholeSummaryDocumentStorageService (#15664)
[AB#4270](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4270)

In our service load tests, we were noticing
`refreshLatestSummary_cancel` events for the AFR endpoints. After
looking at the error stack, we determined that we were fetching the
snapshot, and then trying to read the blob, but the `readBlob` endpoint
does not work for AFR and the blobs weren't cached. Since the blobs are
in the snapshot of the `wholeSummaryDocumentStorageService`. We want to
cache those blobs.

We opted to cache the blobs in the `wholeDocumentSummary/Snapshot` in
the cache.