feat:(fluid-devtools) If the webpage does not have an initialized Devtools, the extension will spin forever (#15322)
## Description

Currently, the extension assumes that it has a corresponding Devtools
instance it can talk to, but this isn't always the case. We need to
update our query logic to not wait indefinitely for data to be returned
from potentially nonexistent Devtools. We can likely accomplish this by
adding a timeout to our root supported-features query - if we don't hear
back within some time period, display a notice to the user that Devtools
could not be found, and what they need to do to enable them.

We should also include a "look again" refresh button just in case there
was actually a communication error, or if the extension was launched
before app initialization.

## Sample