Upload stage timing from perf pipeline to kusto (#15684)
## Description

Today, the pipelines get broken frequently and we do not have a way to
generate alerts whenever it happens. As a proof of concept, this change
adds a stage to the performance benchmark that makes a simple REST call
to the timeline API from dev azure to request information about the
current build id. With that information in hand, we upload to ARIA and,
as a result, we will be able to create monitors and alerts based on the
stage results as well as find out avg duration of each stage.
The idea is to do something similar to other pipelines as the need
arises.

@alexvy86 I believe there is a way to dynamically get the name of all
previous stages and use them to make sure the last one to be run is the
one we've created. For now, I kept it hardcoded but ideally, we should
move in that direction. Thoughts?
Also, notice I'm only using the stages and not the tasks (there are TOO
many tasks and for now I will not upload them).