refactor(devtools): Remove "containerId" concept from Container-level devtools initialization (#15655)
Replaces `containerId` and `containerNickname` with a simple
`containerKey` property which acts as a unique identifier for the
Container when registering it with the Devtools. This key is effectively
the old nickname property - semantically it just needs to be an
identifier to uniquely distinguish Container-level devtools
registrations. It is no longer tied to the Container ID / Document ID.

Also updates the summary view to simply display the key, rather than
attempting to differentiate "Container ID" and "Document ID" concepts..


![image](https://github.com/microsoft/FluidFramework/assets/54606601/d04eaf80-4ed3-48bc-b848-7d0be6452a13)