[r11s] Adding more logs when service restarts (#15691)
## Description

RunService is a metric that is used to track reasons for service
restarts, when those are restarts initiated within the Fluid server
(e.g. an error in Kafka or similar, instead of kubernetes controller
(external to Fluid) kiiling the service).

In AFR, we noticed that occasionally, the RunService metric does not get
emitted/logged, even though the service restart reason was not
kubernetes-initiated. We also ruled out the possibility of our telemetry
pipeline being the issue, as we don't even see RunService console logs
for the affected pod/containers.

In this PR, I add a few more auxiliary log statements that might help
obtain more information about service restarts. I validated this by
forcing errors in two conditions: service runner `start()` and also
service runner `stop()` methods.

I am also using
[serialize-error](https://www.npmjs.com/package/serialize-error) to log
error objects in runner.ts, as it might provide better features than
`util.inspect()`.

## Breaking Changes
None.

## Reviewer Guidance

Note, for future reference: I also noticed that the problem of missing
metrics seem to be related not to our telemetry infra, but instead into
the handling of exceptions. This is not something being addressed in
this PR as it requires additional investigation. I am referring to the
fact that we might have unhandled errors/promise rejections in some
situations - possibly, due to some errors being thrown, while others are
promise rejections - and also due to how we catch those in runner.ts.
The reason I say that: when testing these changes and forcing errors in
service runner `start()` and `stop()`, I noticed the metric is correctly
emitted when the errors were generated with `Promise.reject(error)`. But
the metric would not be available when using `throw error`.