Adds LazyPromise wrapper on DependencyContainer resolver in order to ensure lazy loading of dependencies (#15688)
## Description

This PR solves this
[bug](https://office.visualstudio.com/OC/_workitems/edit/7872230) by
enabling DependencyContainer to work well with LazyPromises, allowing
for delayed loading of expensive dependencies.

## The Problem

Sometimes in loop-app, specific dependencies get mentioned in order to
be passed down the chain, triggering their getter from the
DepenencyContainer, which would end up triggering the LazyPromise
without the use of an await keyword. This invalidates the lazy loading
of these dependencies, causing performance issues in direct nav and
workspace creation flows.

## Testing

Testing was done by reverting loop-app to the [original attempted fix
](https://office.visualstudio.com/OC/_git/office-bohemia/commit/8aa90c748ce4f532a8c8303a719ddcdd8f27cfbf?refName=refs%2Fheads%2Fu%2Fmbeardmore%2Fciq-service-startup-changes)
for [this
issue](https://office.visualstudio.com/OC/_workitems/edit/7799911),
while substituting the ff/synthesize package locally to verify that it
would lazy load correctly. Originally, the bug was not fixed by this
commit and additional work was required to work around the
DC/LazyPromise interaction. Running the tests again with the changes
that are on this PR injected locally, this initial fix does solve the
problem (lazy loading works as expected, only triggering the promise
after an await keyword is used).

A specific example of this pattern happening can be seen in the new unit
tests that were added with this PR as well.

---------

Co-authored-by: Guilherme Bittencourt de Borba <guilhermeb@microsoft.com>
Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>