Performance improvement around token fetching for AFR driver (#15686)
## Description

Performance improvement around token fetching for AFR driver. Based on
learnings from ODSP driver where token fetching does take time, we want
to initiate the token fetch as early as possible for AFR driver. So soon
as the rest wrappers are created, initiate the token fetch so that they
can be used later.