Don't run GC and reset GC data if the GC version is older that the GC version in base snapshot (#15675)
Currently, if the GC versions changes, the GC state is reset, summary is
regenerated and GC runs again with the GC version of the current
container.
This change updates this behavior to the following:
- If the GC version of a container is same or newer than the one in the
base snapshot it loads from, it will reset GC state, regenerate summary
and run GC with the current container's GC version.
- If the GC version of a container is older than the one in the base
snapshot, it will reset GC state and regenerate summary but it will not
run GC anymore.
The reason behind this is that we will typically increment the GC
version when we find a bug in the GC code which can lead to incorrect GC
state. If a container learns that there is a newer GC version out there,
if it continues to run GC with an older version, it may be generating
incorrect GC state and re-introducing bugs fixed by the newer GC
version. So, it should stop running GC to prevent further damage.

[AB#3947](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3947)