Add performance benchmarks for SharedTree's identifier index (#15669)
## Description

These benchmarks measure the following scenarios:

* The time it takes to insert _n_ nodes with an identifier
* The time it takes to lookup _n_ identifiers that exists in the tree
* The time it takes to lookup an identifier that does not exist in the
tree
* The summary size increase caused by having identifiers in the tree

Each of these scenarios is run with different "saturations" of
identifiers in the tree (e.g. a tree where only 5% of its nodes have
identifiers vs. a tree where 75% of its nodes have identifiers) which
determines the value of _n_.

The identifier index is currently un-optimized and has perfomance
characteristics that are probably unacceptable for any real world use,
so the current perf data from these benchmarks is probably also not of
much use. However, they should allow us to measure performance
improvements as the identifier index is optimized (currently a work in
progress).