Add new FieldKind for node identifiers (#15682)
## Description

This PR adds a new FieldKind which is meant to be used exclusively to
store `NodeIdentifiers`. Using a special field kind (rather than simply
putting node identifiers in a `Value` field kind) will make it easier
for the rest of the system to special-case node identifiers. For
example, they can be made immutable at a fundamental level and can be
given special treatment when serializing or during reading.

This also does some related cleanup to node identifiers and the node
identifier index.
* Things named "Identifier" are now named "NodeIdentifier" to reduce
naming collisions with other kinds of identifiers.
* There is now a single function for creating the schema library which
supports node identifiers, and it can be parameterized with different
keys (this will allow different or multiple identifier keys in the
future, if/when SharedTree supports it)
* Minor refactors/cleanup in the identifier index and its tests
* `NodeIdentifier` is now a branded string rather than a number.
* `ISharedTreeView` now exposes a method to generate a unique node
identifier.

## Breaking Changes

* `Identifier` -> `NodeIdentifier`
* Schema symbols for node identifiers are replaced with the result of
`nodeIdentifierSchema()`.
* `NodeIdentifier` is now a string.