Showcasing grouped batches with payloads containing thousands of ops (#15741)
## Description

One of the scenarios that regular compression + chunking does not cover
is when a batch has many small ops. In this case, with compression
operating exclusively on the contents of the ops, the resulting payload
can still go over the 1MB limit due to the accumulated envelope overhead
when the number of ops is large enough.

This test showcases that using the grouped batching features unblocks
this particular scenario.