Decouple Container From Loader (#15700)
## Description
This change removes the coupling between the Container and Loader
classes. The primary mechanism of this change is to split container
constructor input into two object: IContainerLoadProps, and
IContainerCreateProps. CreateProps are always requires to create a
container. LoadProps are additional properties that are specified when
an attached container is loaded.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>