fix(devtools-example): Remove invalid objects from test app, which were breaking data view (#15783)
We were using `MockHandle` (directly, and transitively through
`MockSharedObject`) when setting up our Container. This is fine for a
local-only, transient session. But when attempting to reload the
container (as we can do for the "Shared Container" in the example), the
runtime fails to resolve the handles. This was causing subsequent views
of the Shared Container's data to spin forever awaiting data.

The temporary fix is to remove usages of `MockHandle` from the test app.
In the future, it may be worth adding more legitimate examples of data
the visualizer doesn't understand, but that will require further
investigation.

We also probably need to update our handling of cases like this to give
a better UX experience.