feat: Add `changeset add` command (#15489)
## Problem

Adding a changeset for a change can be done manually, but tooling
improves the experience dramatically. The primary advantages to using a
tool to create a changeset are:

1. Metadata formatting is always done correctly.
2. It can pre-select packages whose contents have been edited when
compared to another git branch.
3. It can generate empty changesets quickly, encouraging people to add
changesets.

The existing changesets tools in
[@changesets/cli](https://github.com/changesets/changesets/blob/main/packages/cli/README.md)
provides these advantages but is insufficient for our needs for the
following reasons:

- **It only supports a single branch.** We use the `next` and `main`
branches regularly to make changeset-worthy changes, and the tool needs
to be able to compare against both of these branches to correctly
populate the list of changed packages.
- **It asks the user to specify the bump type for each package.** We use
the target branch to determine the change type, while the changesets/cli
tool is intended to be used when packages are versioned independently on
a single branch. This makes the tool confusing for Fluid contributors,
because it asks questions that are irrelevant to our repo.
- **There is no way to filter packages.** We have a lot of packages,
many of which are not published, like examples. These unpublished
packages are not usually interesting for changesets, but the existing
tool cannot filter them out.

## Description of this PR

This PR implements a `flub changeset generate` command as a replacement
to changesets/cli. It offers the following features:

- **The list of packages is organized by release group.** This is the
scope that we use when releasing packages.
- **The package selection UX excludes private packages and
example/internal packages by default.** The intent is that by default,
only packages that customers could use should be included. The list of
excluded scopes is hardcoded but could be made configurable in the
future.
- **The list of packages can be filtered by typing.** This UX
improvement makes selecting from the number of packages we have a
bearable experience, even though in most cases the changed packages
should be selected automatically.
- **The package bump type (major/minor) is inferred from the branch
that's being compared with.** This streamlines the process of adding new
changesets and avoids confusing people.


![image](https://github.com/microsoft/FluidFramework/assets/19589/d9de2ab9-4d44-4277-9d02-88b96c829e18)

Fixes
[AB#3967](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3967).

## Testing

To test the command before merge, you can follow these steps:

1. Check out this PR with the GitHub CLI: `gh pr checkout 15489`.
2. `cd build-tools`
3. `pnpm i`
4. `pnpm build:compile`
5. `packages/build-cli/bin/run generate changeset -b main`

## Known issues

1. The "section headers" in the package selection UI are displayed with
a ~~strikethrough~~ because they are disabled. This isn't ideal but
fixing it requires forking the CLI UI, and I don't want to do that yet.

## Remaining work

After this PR is merged, there are still some remaining steps to make
this command usable by developers:

1. We need to publish a new version of build-tools that contains this
command.
6. Then we need to integrate that version of build-tools into the client
release group.
7. Add scripts that call the new command.
8. Update documentation in wiki.