Interval undo redo (#15778)
## Description

This is the combined PR for revertibles (i.e. undo-redo) for intervals.
This includes all edits directly on intervals as well as string edits
that affect them. Users should use only this code and not merge tree
revertibles directly.

The individual changes are:

- https://github.com/jzaffiro/FluidFramework/pull/1
- https://github.com/jzaffiro/FluidFramework/pull/3
- https://github.com/jzaffiro/FluidFramework/pull/4
- https://github.com/jzaffiro/FluidFramework/pull/5
- https://github.com/jzaffiro/FluidFramework/pull/6
- https://github.com/jzaffiro/FluidFramework/pull/7
- https://github.com/jzaffiro/FluidFramework/pull/8


[AB#3578](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3578)

Note: a test farm is still in-progress, after which we'll consider this
customer-ready.

## Reviewer Guidance

Although all these changes have been reviewed, comments on any aspect
are welcome. Also looking for feedback on the integrated totality of
changes.

---------

Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>