refactor(devtools-core): Make message handlers async (#15792)
Updates our message handler infra to make handlers async. The current
motivation for this is to allow the data visualization code to run to
completion before we report whether or not the message was handled.

### Breaking Change

This PR changes the message handler interface, which is exported and
used by the view.