Add support for logging flat objects (e.g. details) to reduce number of unique columns in telemetry (#15827)
## Description

Precursors:
* PR https://github.com/microsoft/FluidFramework/pull/12863 moved the
typings for our internal loggers to telemetry-utils package and added
support for arrays.
* PR #15667 upgraded the client release group to use the new types

This PR updates that type further to add support for flat objects. This
has become a more common practice to consolidate event-specific
properties, e.g. in a `details` object. Today instrumentation does the
JSON.stringify itself, but this is not ideal. It's proven the utility of
logging properties like this, so this PR adds first-class support for
this.

## Breaking Changes

Yes, this is a breaking change. If a logger implements the existing
logger contracts based on the existing type, it will not necessarily be
equipped to handle a flat object.

In practice, there shouldn't be any follow-up required. These types are
only used for our internal logging contracts, which are always fulfilled
by a class derived from `TelemetryLogger`, which actually handles this
case just fine without any changes.

The API surface we expect partners to use always speaks in
`ITelemetryBaseLogger` which is unchanged here.