offline: fix offline load 404 bug (#15831)
Fix
[AB#4399](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4399)

Fix 404 errors when stashed ops enabled in stress tests and re-enable
stashed ops in stress test config.

ContainerStorageAdapter would previously avoid saving the ".blobs" tree
from the snapshot since this was causing it to save all attachment blobs
since they are listed in this tree and are indistinguishable from other
snapshot blobs. This caused problems though since the BlobManager
redirect table is also saved in the ".blobs" tree, so a 404 would result
when loading from a stash blob if the snapshot in the stash blob has
been deleted from storage.

A consequence of this change is that all attachment blobs will be saved
in the stash blob.