fix(devtools-browser-extension): Fix globals stubbing infra with regards to the `window` object (#15848)
#15829 Added new globals stubbing infra for use with the `window` object
for testing the Content script. It inadvertently broke the extension
build (verified when you attempt to install it in the browser), due to
the `window` object not being defined in the BackgroundScript context.
This is expected per Google's documentation.

This PR adds handling to support the case where `window` is not defined.