build(devtools-browser-extension): Update manifest "name" property to reflect preview status of the tooling (#15904)
Bumps the version to 0.0.2 so we can publish the update to the stores.
Also adds some documentation to make changes a bit easier in the future.

Note: Added a temporary changelog file. I'm hoping to change the name to
"CHANGELOG", but for now that conflicts with a log that is automatically
generated by our tooling. I am investigating how to opt out of that
behavior.