Fix lts (#15909)
This PR fixes the LTS branch so that it builds correctly for the azure
release group. Fixes
[AB#4214](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4214).

## Summary of changes

- Pinned build-tools to 0.10.0, the last version that fully supported
npm.
- Version 0.11.0 was skipped, and version 0.12 removed the `flub
generate typetests` entrypoint.
- The alternative is disabling the type tests and using a more recent
version of build-tools with the fixes in #15912. We can upgrade
build-tools separately later if needed.
- Copied the code changes for azure 1.2.0 release from main to lts.
Basically this was just copying the contents of the azure folder in main
to lts and cleaning up a few things.
- Updated build-common to 1.2.0. (this wasn't needed, but it gets us in
sync with main for default prettier settings and tsconfig)
- Reformatted code using prettier. My rationale is that it will make it
easier to port changes between lts and main.
- Copied the prettierignore file from main to lts. I only copied the
root one, since lts is only used for azure and client (not server).
- Moved azure-local-service into the azure release group. My goal was to
make it as much like main as reasonable. **This causes the
azure-local-service pipeline to fail. Once merged we can disable that
pipeline.**