Consume build common with target ES2020 (#15941)
## Description

Item: https://dev.azure.com/fluidframework/internal/_workitems/edit/2772

Bump build common dependency with target ES2020 in tsconfig. We also
needed to upgrade versions of "puppeteer" and "jest-puppeteer" as part
of this.
Puppeteer is tied to a specific version of Chromium, and the version we
were using were corresponding to an older browser that doesn't support
all the es2020 constructs. So we needed to upgrade puppeteer.
Jest-puppeteer also needed to be upgraded as puppeteer was peer
dependency and new version of puppeteer was not supported by present
version of jest-puppeteer.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>