[GC Stress] Add multi client collaboration for attachment blobs (#15951)
## Description
Added multi client collab for attachment blobs. When a client references
or unreferenced an attachment blob, one partner client will run or stop
activity on its local copy respectively. This should give us coverage
where multiple clients retrieve the same attachment blob, for e.g., an
image.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.


[AB#2750](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2750)