Improve Detached Container Tests (#15977)
During prototyping I ran into an issue with these tests. Specifically,
they took a dependency on the fact the attaching state would by
synchronously reached when attach is called. This is specifically not
true in the case of blobs today, as blobs are asynchronously uploaded
before attachment begins. So making a light improvement to these tests
to better capture the pre-conditions these tests need, and avoid needing
to debug them in the future.

Co-authored-by: Tony Murphy <anthonm@microsoft.com>